requirements.txt
imageio==2.4.1
librosa==0.9.2
hydra-core
moviepy
numpy
torch
tqdm==4.64.1
transformers==4.33.2
